# docs/openapi_spec.yaml
openapi: "3.0.3"
info:
  title: Supply Chain Risk Intelligence API
  description: |
    24-dimensional passive supply-chain risk detection platform (SCRI).
    Aggregates 30+ open data sources to provide real-time risk scoring,
    BOM analysis, cost-impact estimation, screening, analytics, and
    forecasting capabilities.
  version: "0.9.0"
  contact:
    name: SCRI Platform Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:8000
    description: Local development server
...
paths:
...
  /api/v1/bom/analyze:
    post:
      tags: [BOM Analysis]
      summary: Analyze BOM risk
      description: Perform risk analysis on a Bill of Materials with optional Tier-2 inference.
      operationId: analyzeBom
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BOMAnalysisRequest"
      responses:
        "200":
          description: BOM risk analysis
          content:
            application/json:
              schema:
                type: object
        "502":
          $ref: "#/components/responses/UpstreamError"
  /api/v1/bom/infer-supply-chain:
    post:
      tags: [BOM Analysis]
      summary: Infer Tier-2/3 supply chain
      description: Estimate hidden supply chain tiers beyond Tier-1 suppliers.
      operationId: inferSupplyChain
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TierInferenceRequest"
      responses:
        "200":
          description: Inferred supply chain
          content:
            application/json:
              schema:
                type: object
        "502":
          $ref: "#/components/responses/UpstreamError"
  /api/v1/bom/hidden-risk:
    post:
      tags: [BOM Analysis]
      summary: Hidden risk exposure
      description: Analyze hidden risk exposure from inferred sub-tier suppliers.
      operationId: getHiddenRisk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HiddenRiskRequest"
      responses:
        "200":
          description: Hidden risk analysis
          content:
            application/json:
              schema:
                type: object
        "502":
          $ref: "#/components/responses/UpstreamError"
  /api/v1/bom/import-csv:
    post:
      tags: [BOM Analysis]
      summary: Import BOM from CSV
      description: Import a BOM from CSV data and run risk analysis.
      operationId: importBomCsv
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BOMImportRequest"
      responses:
        "200":
          description: BOM analysis result
          content:
            application/json:
              schema:
                type: object
        "400":
          description: Invalid CSV data
        "502":
          $ref: "#/components/responses/UpstreamError"
  /api/v1/bom/product-risk:
    get:
      tags: [BOM Analysis]
      summary: Product risk from BOM file
      description: Calculate product risk from a BOM JSON file path.
      operationId: getProductRisk
      parameters:
        - name: product_name
          in: query
          schema:
            type: string
            default: Product
        - name: bom_file
          in: query
          schema:
            type: string
          description: Path to BOM JSON file
      responses:
        "200":
          description: Product risk comparison
          content:
            application/json:
              schema:
                type: object
        "502":
          $ref: "#/components/responses/UpstreamError"
...
components:
  schemas:
...
    BOMAnalysisRequest:
      type: object
      required: [bom]
      properties:
        bom:
          type: array
          items:
            type: object
            properties:
              part_id:
                type: string
              part_name:
                type: string
              supplier_name:
                type: string
              supplier_country:
                type: string
              material:
                type: string
              hs_code:
                type: string
              tier:
                type: integer
              quantity:
                type: number
              unit_cost_usd:
                type: number
              is_critical:
                type: boolean
        product_name:
          type: string
          default: Product
        include_tier2_inference:
          type: boolean
          default: false
    TierInferenceRequest:
      type: object
      required: [tier1_country, hs_code]
      properties:
        tier1_country:
          type: string
        hs_code:
          type: string
        material:
          type: string
          default: ""
        max_depth:
          type: integer
          default: 3
    HiddenRiskRequest:
      type: object
      required: [tier1_country, materials]
      properties:
        tier1_country:
          type: string
        materials:
          type: array
          items:
            type: object
            properties:
              hs_code:
                type: string
              material:
                type: string
    BOMImportRequest:
      type: object
      required: [csv_data]
      properties:
        csv_data:
          type: string
        product_name:
          type: string
          default: Product
        include_tier2_inference:
          type: boolean
          default: false
...
  responses:
...
    UpstreamError:
      description: Upstream data source error
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string